Complete restructure of the project
