--- reference/src/lib/gost34112018_private.h
@@ -1,310 +1,0 @@
-// Copyright 2025, Anufriev Ilia, anufriewwi@rambler.ru
-// SPDX-License-Identifier: BSD-3-Clause-No-Military-License OR GPL-3.0-or-later
-
-#ifndef __GOST34112018_PRIVATE_H__
-#define __GOST34112018_PRIVATE_H__
-
-#ifdef DEBUG
-    #include "stdio.h"
-    #define log_d(__fmt, ...) \
-        printf("GOST34112018 [DEBUG] %s: " __fmt "\n", __func__,##__VA_ARGS__)
-#else
-    #define log_d(__fmd, ...) ;
-#endif // DEBUG
-
-enum GOST_34112018_CONSTANTS
-{
-    BYTE_SIZE     = (1 * 8), // bits in byte
-    WORD_SIZE     = (2 * 8), // bits in word (16-bit integer)
-    DWORD_SIZE    = (4 * 8), // bits in double-word (32-bit integer)
-    QWORD_SIZE    = (8 * 8), // bits in quad-word (64-bit integer)
-
-    VEC512_BYTES  = 512 / BYTE_SIZE,
-    VEC512_WORDS  = 512 / WORD_SIZE,
-    VEC512_DWORDS = 512 / DWORD_SIZE,
-    VEC512_QWORDS = 512 / QWORD_SIZE,
-
-    VEC512_SIZE   = VEC512_BYTES,
-
-    BLOCK_SIZE    = 512 / BYTE_SIZE,
-};
-
-typedef unsigned long long   GostU64;
-typedef unsigned int         GostU32;
-typedef unsigned short       GostU16;
-typedef unsigned char        GostU8;
-
-typedef long long            GostI64;
-typedef int                  GostI32;
-typedef short                GostI16;
-typedef char                 GostI8;
-
-typedef enum { false, true } GostBool;
-
-/**
-    @brief       Big-endian unsigned 512-bit number.
- */
-union Vec512 {
-    GostU8  bytes  [VEC512_BYTES];
-    GostU16 words  [VEC512_WORDS];
-    GostU32 dwords [VEC512_DWORDS];
-    GostU64 qwords [VEC512_QWORDS];
-};
-
-extern const union Vec512 INIT_VECTOR_256;
-extern const union Vec512 INIT_VECTOR_512;
-
-/**
-    @brief      Context of the algorithm, as described in ch. 8.1 of the Standard.
- */
-struct GOST34112018_Context {
-    union Vec512 h;
-    union Vec512 N;
-    union Vec512 sigma;
-};
-
-/**
-    @brief      S-box, as defined in the chapter 5.1 of The Standard. It is used for
-                S-transformation.
- */
-GostU8 PI[256] = {
-    252, 238, 221,  17, 207, 110,  49,  22,
-    251, 196, 250, 218,  35, 197,   4,  77,
-    233, 119, 240, 219, 147,  46, 153, 186,
-     23,  54, 241, 187,  20, 205,  95, 193,
-    249,  24, 101,  90, 226,  92, 239,  33,
-    129,  28,  60,  66, 139,   1, 142,  79,
-      5, 132,   2, 174, 227, 106, 143, 160,
-      6,  11, 237, 152, 127, 212, 211,  31,
-    235,  52,  44,  81, 234, 200,  72, 171,
-    242,  42, 104, 162, 253,  58, 206, 204,
-    181, 112,  14,  86,   8,  12, 118,  18,
-    191, 114,  19,  71, 156, 183,  93, 135,
-     21, 161, 150,  41,  16, 123, 154, 199,
-    243, 145, 120, 111, 157, 158, 178, 177,
-     50, 117,  25,  61, 255,  53, 138, 126,
-    109,  84, 198, 128, 195, 189,  13,  87,
-    223, 245,  36, 169,  62, 168,  67, 201,
-    215, 121, 214, 246, 124,  34, 185,   3,
-    224,  15, 236, 222, 122, 148, 176, 188,
-    220, 232,  40,  80,  78,  51,  10,  74,
-    167, 151,  96, 115,  30,   0,  98,  68,
-     26, 184,  56, 130, 100, 159,  38,  65,
-    173,  69,  70, 146,  39,  94,  85,  47,
-    140, 163, 165, 125, 105, 213, 149,  59,
-      7,  88, 179,  64, 134, 172,  29, 247,
-     48,  55, 107, 228, 136, 217, 231, 137,
-    225,  27, 131,  73,  76,  63, 248, 254,
-    141,  83, 170, 144, 202, 216, 133,  97,
-     32, 113, 103, 164,  45,  43,   9,  91,
-    203, 155,  37, 208, 190, 229, 108,  82,
-     89, 166, 116, 210, 230, 244, 180, 192,
-    209, 102, 175, 194,  57,  75,  99, 182
-};
-
-/**
-    @brief      A byte shuffle table, as defined in the ch. 5.3 of The Standard. It is
-                used for the P transformation.
- */
-GostU8 TAU[64] = {
-    0,  8, 16, 24, 32, 40, 48, 56,
-    1,  9, 17, 25, 33, 41, 49, 57,
-    2, 10, 18, 26, 34, 42, 50, 58,
-    3, 11, 19, 27, 35, 43, 51, 59,
-    4, 12, 20, 28, 36, 44, 52, 60,
-    5, 13, 21, 29, 37, 45, 53, 61,
-    6, 14, 22, 30, 38, 46, 54, 62,
-    7, 15, 23, 31, 39, 47, 55, 63
-};
-
-enum
-{
-    A_SIZE = 64,
-    C_SIZE = 12,
-};
-
-GostU64 A[64] = {
-    0x8e20faa72ba0b470, 0x47107ddd9b505a38, 0xad08b0e0c3282d1c, 0xd8045870ef14980e,
-    0x6c022c38f90a4c07, 0x3601161cf205268d, 0x1b8e0b0e798c13c8, 0x83478b07b2468764,
-    0xa011d380818e8f40, 0x5086e740ce47c920, 0x2843fd2067adea10, 0x14aff010bdd87508,
-    0x0ad97808d06cb404, 0x05e23c0468365a02, 0x8c711e02341b2d01, 0x46b60f011a83988e,
-    0x90dab52a387ae76f, 0x486dd4151c3dfdb9, 0x24b86a840e90f0d2, 0x125c354207487869,
-    0x092e94218d243cba, 0x8a174a9ec8121e5d, 0x4585254f64090fa0, 0xaccc9ca9328a8950,
-    0x9d4df05d5f661451, 0xc0a878a0a1330aa6, 0x60543c50de970553, 0x302a1e286fc58ca7,
-    0x18150f14b9ec46dd, 0x0c84890ad27623e0, 0x0642ca05693b9f70, 0x0321658cba93c138,
-    0x86275df09ce8aaa8, 0x439da0784e745554, 0xafc0503c273aa42a, 0xd960281e9d1d5215,
-    0xe230140fc0802984, 0x71180a8960409a42, 0xb60c05ca30204d21, 0x5b068c651810a89e,
-    0x456c34887a3805b9, 0xac361a443d1c8cd2, 0x561b0d22900e4669, 0x2b838811480723ba,
-    0x9bcf4486248d9f5d, 0xc3e9224312c8c1a0, 0xeffa11af0964ee50, 0xf97d86d98a327728,
-    0xe4fa2054a80b329c, 0x727d102a548b194e, 0x39b008152acb8227, 0x9258048415eb419d,
-    0x492c024284fbaec0, 0xaa16012142f35760, 0x550b8e9e21f7a530, 0xa48b474f9ef5dc18,
-    0x70a6a56e2440598e, 0x3853dc371220a247, 0x1ca76e95091051ad, 0x0edd37c48a08a6d8,
-    0x07e095624504536c, 0x8d70c431ac02a736, 0xc83862965601dd1b, 0x641c314b2b8ee083
-};
-
-union Vec512 C1 = {
-    .bytes = {
-        0xb1, 0x08, 0x5b, 0xda, 0x1e, 0xca, 0xda, 0xe9,
-        0xeb, 0xcb, 0x2f, 0x81, 0xc0, 0x65, 0x7c, 0x1f,
-        0x2f, 0x6a, 0x76, 0x43, 0x2e, 0x45, 0xd0, 0x16,
-        0x71, 0x4e, 0xb8, 0x8d, 0x75, 0x85, 0xc4, 0xfc,
-        0x4b, 0x7c, 0xe0, 0x91, 0x92, 0x67, 0x69, 0x01,
-        0xa2, 0x42, 0x2a, 0x08, 0xa4, 0x60, 0xd3, 0x15,
-        0x05, 0x76, 0x74, 0x36, 0xcc, 0x74, 0x4d, 0x23,
-        0xdd, 0x80, 0x65, 0x59, 0xf2, 0xa6, 0x45, 0x07
-    }
-};
-
-union Vec512 C2 = {
-    .bytes = {
-        0x6f, 0xa3, 0xb5, 0x8a, 0xa9, 0x9d, 0x2f, 0x1a,
-        0x4f, 0xe3, 0x9d, 0x46, 0x0f, 0x70, 0xb5, 0xd7,
-        0xf3, 0xfe, 0xea, 0x72, 0x0a, 0x23, 0x2b, 0x98,
-        0x61, 0xd5, 0x5e, 0x0f, 0x16, 0xb5, 0x01, 0x31,
-        0x9a, 0xb5, 0x17, 0x6b, 0x12, 0xd6, 0x99, 0x58,
-        0x5c, 0xb5, 0x61, 0xc2, 0xdb, 0x0a, 0xa7, 0xca,
-        0x55, 0xdd, 0xa2, 0x1b, 0xd7, 0xcb, 0xcd, 0x56,
-        0xe6, 0x79, 0x04, 0x70, 0x21, 0xb1, 0x9b, 0xb7
-    }
-};
-
-union Vec512 C3 = {
-    .bytes = {
-        0xf5, 0x74, 0xdc, 0xac, 0x2b, 0xce, 0x2f, 0xc7,
-        0x0a, 0x39, 0xfc, 0x28, 0x6a, 0x3d, 0x84, 0x35,
-        0x06, 0xf1, 0x5e, 0x5f, 0x52, 0x9c, 0x1f, 0x8b,
-        0xf2, 0xea, 0x75, 0x14, 0xb1, 0x29, 0x7b, 0x7b,
-        0xd3, 0xe2, 0x0f, 0xe4, 0x90, 0x35, 0x9e, 0xb1,
-        0xc1, 0xc9, 0x3a, 0x37, 0x60, 0x62, 0xdb, 0x09,
-        0xc2, 0xb6, 0xf4, 0x43, 0x86, 0x7a, 0xdb, 0x31,
-        0x99, 0x1e, 0x96, 0xf5, 0x0a, 0xba, 0x0a, 0xb2
-    }
-};
-
-union Vec512 C4 = {
-    .bytes = {
-        0xef, 0x1f, 0xdf, 0xb3, 0xe8, 0x15, 0x66, 0xd2,
-        0xf9, 0x48, 0xe1, 0xa0, 0x5d, 0x71, 0xe4, 0xdd,
-        0x48, 0x8e, 0x85, 0x7e, 0x33, 0x5c, 0x3c, 0x7d,
-        0x9d, 0x72, 0x1c, 0xad, 0x68, 0x5e, 0x35, 0x3f,
-        0xa9, 0xd7, 0x2c, 0x82, 0xed, 0x03, 0xd6, 0x75,
-        0xd8, 0xb7, 0x13, 0x33, 0x93, 0x52, 0x03, 0xbe,
-        0x34, 0x53, 0xea, 0xa1, 0x93, 0xe8, 0x37, 0xf1,
-        0x22, 0x0c, 0xbe, 0xbc, 0x84, 0xe3, 0xd1, 0x2e
-    }
-};
-
-union Vec512 C5 = {
-    .bytes = {
-        0x4b, 0xea, 0x6b, 0xac, 0xad, 0x47, 0x47, 0x99,
-        0x9a, 0x3f, 0x41, 0x0c, 0x6c, 0xa9, 0x23, 0x63,
-        0x7f, 0x15, 0x1c, 0x1f, 0x16, 0x86, 0x10, 0x4a,
-        0x35, 0x9e, 0x35, 0xd7, 0x80, 0x0f, 0xff, 0xbd,
-        0xbf, 0xcd, 0x17, 0x47, 0x25, 0x3a, 0xf5, 0xa3,
-        0xdf, 0xff, 0x00, 0xb7, 0x23, 0x27, 0x1a, 0x16,
-        0x7a, 0x56, 0xa2, 0x7e, 0xa9, 0xea, 0x63, 0xf5,
-        0x60, 0x17, 0x58, 0xfd, 0x7c, 0x6c, 0xfe, 0x57
-    },
-};
-
-union Vec512 C6 = {
-    .bytes = {
-        0xae, 0x4f, 0xae, 0xae, 0x1d, 0x3a, 0xd3, 0xd9,
-        0x6f, 0xa4, 0xc3, 0x3b, 0x7a, 0x30, 0x39, 0xc0,
-        0x2d, 0x66, 0xc4, 0xf9, 0x51, 0x42, 0xa4, 0x6c,
-        0x18, 0x7f, 0x9a, 0xb4, 0x9a, 0xf0, 0x8e, 0xc6,
-        0xcf, 0xfa, 0xa6, 0xb7, 0x1c, 0x9a, 0xb7, 0xb4,
-        0x0a, 0xf2, 0x1f, 0x66, 0xc2, 0xbe, 0xc6, 0xb6,
-        0xbf, 0x71, 0xc5, 0x72, 0x36, 0x90, 0x4f, 0x35,
-        0xfa, 0x68, 0x40, 0x7a, 0x46, 0x64, 0x7d, 0x6e
-    }
-};
-
-union Vec512 C7 = {
-    .bytes = {
-        0xf4, 0xc7, 0x0e, 0x16, 0xee, 0xaa, 0xc5, 0xec,
-        0x51, 0xac, 0x86, 0xfe, 0xbf, 0x24, 0x09, 0x54,
-        0x39, 0x9e, 0xc6, 0xc7, 0xe6, 0xbf, 0x87, 0xc9,
-        0xd3, 0x47, 0x3e, 0x33, 0x19, 0x7a, 0x93, 0xc9,
-        0x09, 0x92, 0xab, 0xc5, 0x2d, 0x82, 0x2c, 0x37,
-        0x06, 0x47, 0x69, 0x83, 0x28, 0x4a, 0x05, 0x04,
-        0x35, 0x17, 0x45, 0x4c, 0xa2, 0x3c, 0x4a, 0xf3,
-        0x88, 0x86, 0x56, 0x4d, 0x3a, 0x14, 0xd4, 0x93
-    }
-};
-
-union Vec512 C8 = {
-    .bytes = {
-        0x9b, 0x1f, 0x5b, 0x42, 0x4d, 0x93, 0xc9, 0xa7,
-        0x03, 0xe7, 0xaa, 0x02, 0x0c, 0x6e, 0x41, 0x41,
-        0x4e, 0xb7, 0xf8, 0x71, 0x9c, 0x36, 0xde, 0x1e,
-        0x89, 0xb4, 0x44, 0x3b, 0x4d, 0xdb, 0xc4, 0x9a,
-        0xf4, 0x89, 0x2b, 0xcb, 0x92, 0x9b, 0x06, 0x90,
-        0x69, 0xd1, 0x8d, 0x2b, 0xd1, 0xa5, 0xc4, 0x2f,
-        0x36, 0xac, 0xc2, 0x35, 0x59, 0x51, 0xa8, 0xd9,
-        0xa4, 0x7f, 0x0d, 0xd4, 0xbf, 0x02, 0xe7, 0x1e
-    }
-};
-
-union Vec512 C9 = {
-    .bytes = {
-        0x37, 0x8f, 0x5a, 0x54, 0x16, 0x31, 0x22, 0x9b,
-        0x94, 0x4c, 0x9a, 0xd8, 0xec, 0x16, 0x5f, 0xde,
-        0x3a, 0x7d, 0x3a, 0x1b, 0x25, 0x89, 0x42, 0x24,
-        0x3c, 0xd9, 0x55, 0xb7, 0xe0, 0x0d, 0x09, 0x84,
-        0x80, 0x0a, 0x44, 0x0b, 0xdb, 0xb2, 0xce, 0xb1,
-        0x7b, 0x2b, 0x8a, 0x9a, 0xa6, 0x07, 0x9c, 0x54,
-        0x0e, 0x38, 0xdc, 0x92, 0xcb, 0x1f, 0x2a, 0x60,
-        0x72, 0x61, 0x44, 0x51, 0x83, 0x23, 0x5a, 0xdb
-    }
-};
-
-union Vec512 C10 = {
-    .bytes = {
-        0xab, 0xbe, 0xde, 0xa6, 0x80, 0x05, 0x6f, 0x52,
-        0x38, 0x2a, 0xe5, 0x48, 0xb2, 0xe4, 0xf3, 0xf3,
-        0x89, 0x41, 0xe7, 0x1c, 0xff, 0x8a, 0x78, 0xdb,
-        0x1f, 0xff, 0xe1, 0x8a, 0x1b, 0x33, 0x61, 0x03,
-        0x9f, 0xe7, 0x67, 0x02, 0xaf, 0x69, 0x33, 0x4b,
-        0x7a, 0x1e, 0x6c, 0x30, 0x3b, 0x76, 0x52, 0xf4,
-        0x36, 0x98, 0xfa, 0xd1, 0x15, 0x3b, 0xb6, 0xc3,
-        0x74, 0xb4, 0xc7, 0xfb, 0x98, 0x45, 0x9c, 0xed
-    }
-};
-
-union Vec512 C11 = {
-    .bytes = {
-        0x7b, 0xcd, 0x9e, 0xd0, 0xef, 0xc8, 0x89, 0xfb,
-        0x30, 0x02, 0xc6, 0xcd, 0x63, 0x5a, 0xfe, 0x94,
-        0xd8, 0xfa, 0x6b, 0xbb, 0xeb, 0xab, 0x07, 0x61,
-        0x20, 0x01, 0x80, 0x21, 0x14, 0x84, 0x66, 0x79,
-        0x8a, 0x1d, 0x71, 0xef, 0xea, 0x48, 0xb9, 0xca,
-        0xef, 0xba, 0xcd, 0x1d, 0x7d, 0x47, 0x6e, 0x98,
-        0xde, 0xa2, 0x59, 0x4a, 0xc0, 0x6f, 0xd8, 0x5d,
-        0x6b, 0xca, 0xa4, 0xcd, 0x81, 0xf3, 0x2d, 0x1b
-    }
-};
-
-union Vec512 C12 = {
-    .bytes = {
-        0x37, 0x8e, 0xe7, 0x67, 0xf1, 0x16, 0x31, 0xba,
-        0xd2, 0x13, 0x80, 0xb0, 0x04, 0x49, 0xb1, 0x7a,
-        0xcd, 0xa4, 0x3c, 0x32, 0xbc, 0xdf, 0x1d, 0x77,
-        0xf8, 0x20, 0x12, 0xd4, 0x30, 0x21, 0x9f, 0x9b,
-        0x5d, 0x80, 0xef, 0x9d, 0x18, 0x91, 0xcc, 0x86,
-        0xe7, 0x1d, 0xa4, 0xaa, 0x88, 0xe1, 0x28, 0x52,
-        0xfa, 0xf4, 0x17, 0xd5, 0xd9, 0xb2, 0x1b, 0x99,
-        0x48, 0xbc, 0x92, 0x4a, 0xf1, 0x1b, 0xd7, 0x20
-    }
-};
-
-/**
-    @brief      Iteration constants, that are used in the encryption function of the
-                algorithm.
-    @note       All constants are Big-Endian unsigned 512-bit numbers, which means that,
-                for example, C1.bytes[0] is MSB and C1.bytes[63] is LSB.
- */
-union Vec512 *C[] = { &C1, &C2, &C3,  &C4,  &C5,  &C6,
-                      &C7, &C8, &C9, &C10, &C11, &C12 };
-
-#endif // __GOST34112018_PRIVATE_H__
